refactor: rename cron.providers.ts to cron.provider.ts
